# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_ftp.BandwidthConfig:
    properties:
      download_kbps:
        description: 下载速率限制 (KB/s), 0 表示无限制
        type: integer
      enabled:
        type: boolean
      per_user:
        description: 是否按用户独立限制
        type: boolean
      upload_kbps:
        description: 上传速率限制 (KB/s), 0 表示无限制
        type: integer
    type: object
  internal_ftp.Config:
    properties:
      allow_anonymous:
        description: 允许匿名登录
        type: boolean
      bandwidth_limit:
        allOf:
        - $ref: '#/definitions/internal_ftp.BandwidthConfig'
        description: 带宽限制
      enabled:
        type: boolean
      max_connections:
        description: 最大连接数
        type: integer
      pasv_host:
        description: 被动模式对外 IP
        type: string
      pasv_port_end:
        description: 被动模式端口范围结束
        type: integer
      pasv_port_start:
        description: 被动模式端口范围起始
        type: integer
      port:
        description: FTP 端口 (默认 21)
        type: integer
      root_path:
        description: 根目录
        type: string
      virtual_dirs:
        additionalProperties:
          type: string
        description: '虚拟目录映射: 虚拟路径 -> 实际路径'
        type: object
    type: object
...
  internal_nfs.Client:
    properties:
      host:
        type: string
      options:
        items:
          type: string
        type: array
    type: object
  internal_nfs.Export:
    properties:
      clients:
        items:
          $ref: '#/definitions/internal_nfs.Client'
        type: array
      comment:
        type: string
      fsid:
        type: integer
      options:
        $ref: '#/definitions/internal_nfs.ExportOptions'
      path:
        type: string
    type: object
  internal_nfs.ExportOptions:
    properties:
      async:
        type: boolean
      no_root_squash:
        type: boolean
      ro:
        type: boolean
      rw:
        type: boolean
      secure:
        type: boolean
      subtree_check:
        type: boolean
    type: object
  internal_nfs.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@nas-os.io
    name: NAS-OS 团队
    url: https://github.com/nas-os/nas-os
  description: |-
    NAS-OS 是一个现代化的网络存储操作系统，提供卷管理、用户管理、共享管理、网络配置等功能。

    ## 功能模块
    - **卷管理**: Btrfs 卷创建、快照、RAID 配置
    - **用户管理**: 用户/组管理、认证授权
    - **共享管理**: SMB/NFS 共享配置
    - **网络管理**: 网络接口、DDNS、防火墙、端口转发
    - **Docker 管理**: 容器、镜像、应用商店
    - **插件系统**: 可扩展的插件架构
    - **配额管理**: 存储配额控制
    - **性能监控**: 系统性能监控与报告
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: NAS-OS API
...
paths:
...
  /ftp/config:
    get:
      consumes:
      - application/json
      description: 获取 FTP 服务器配置
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 获取 FTP 配置
      tags:
      - ftp
    put:
      consumes:
      - application/json
      description: 更新 FTP 服务器配置
      parameters:
      - description: 配置
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/internal_ftp.Config'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      summary: 更新 FTP 配置
      tags:
      - ftp
...
  /nfs/exports:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_nfs.Response'
      summary: 列出所有NFS导出
      tags:
      - NFS
    post:
      consumes:
      - application/json
      parameters:
      - description: 导出配置
        in: body
        name: export
        required: true
        schema:
          $ref: '#/definitions/internal_nfs.Export'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_nfs.Response'
      summary: 创建NFS导出
      tags:
      - NFS
  /nfs/exports/{path}:
    delete:
      parameters:
      - description: 导出路径
        in: path
        name: path
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_nfs.Response'
      summary: 删除NFS导出
      tags:
      - NFS
    get:
      parameters:
      - description: 导出路径
        in: path
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_nfs.Response'
      summary: 获取NFS导出详情
      tags:
      - NFS
    put:
      consumes:
      - application/json
      parameters:
      - description: 导出路径
        in: path
        name: path
        required: true
        type: string
      - description: 导出配置
        in: body
        name: export
        required: true
        schema:
          $ref: '#/definitions/internal_nfs.Export'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_nfs.Response'
      summary: 更新NFS导出
      tags:
      - NFS
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API 密钥认证
    in: query
    name: token
    type: apiKey
  BearerAuth:
    description: JWT 认证令牌，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...